+++ src/track/traits.rs
@@ -1269,6 +1269,7 @@ D\tL1\thuman\t95
                     to,
                     low,
                     high,
+                    ..
                 } => Some((label, from, to, low, high)),
                 _ => None,
             })